+++ h3ron-graph/CHANGES.md
@@ -16,6 +16,7 @@ python extension.
 * added `WithinWeightThreshold` and `WithinWeightThresholdMany` traits.
 * add `Path::cells()`.
 * Implemented `h3ron::to_geo::ToLineString` for `LongEdge`.
+* Add `Path::length_m()`.
 
 ### Changed